refactor the EAMode class into an enum and assembly parameter
