+++ src/easier68k/core/models/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['list_file']
+__all__ = ['list_file', 'assembly_parameter']
