+++ src/easier68k/core/util/parsing.py
@@ -1,5 +1,103 @@
 # Parsing utils
 
+from ..enum.ea_mode import EAMode
+from ..models.assembly_parameter import AssemblyParameter
+
+def parse_assembly_parameter(addr: str) -> AssemblyParameter:
+    """
+    Parses an effective addressing mode (such as D0, (A1), #$01)
+    and makes a new AssemblyParameter
+
+    >>> parse_assembly_parameter('D')
+    Traceback (most recent call last):
+    ...
+    AssertionError
+
+    >>> str(parse_assembly_parameter('D3'))
+    'Mode: 0, Data: 3'
+
+    >>> str(parse_assembly_parameter('A6'))
+    'Mode: 1, Data: 6'
+
+    >>> str(parse_assembly_parameter('(A4)'))
+    'Mode: 2, Data: 4'
+
+    >>> str(parse_assembly_parameter('(A2)+'))
+    'Mode: 3, Data: 2'
+
+    >>> str(parse_assembly_parameter('(A2)-'))  # Invalid, can't do "post-decrement"
+    Traceback (most recent call last):
+    ...
+    AssertionError
+
+    >>> str(parse_assembly_parameter('($45).W'))
+    "Mode: 7, Data: bytearray(b'E')"
+
+    >>> str(parse_assembly_parameter('(%01010111).L'))
+    "Mode: 6, Data: bytearray(b'W')"
+
+    >>> str(parse_assembly_parameter('#$FF'))
+    "Mode: 5, Data: bytearray(b'\\\\xff')"
+
+    >>> str(parse_assembly_parameter('-(A2)'))
+    'Mode: 4, Data: 2'
+    """
+    assert len(addr) >= 2
+
+    if addr[0] == 'D':
+        assert len(addr) == 2
+        assert 0 <= int(addr[1]) <= 7
+        return AssemblyParameter(EAMode.DRD, int(addr[1]))
+    if addr[0] == 'A':
+        assert len(addr) == 2
+        assert 0 <= int(addr[1]) <= 7
+        return AssemblyParameter(EAMode.ARD, int(addr[1]))
+    if addr[0] == '(':  # ARI, ARIPI, ALA, or AWA
+        # Parse the inside of the parentheses
+        nested = ""
+        found_paren = False
+
+        i = 1
+        while i < len(addr):
+            if addr[i] == ')':
+                found_paren = True
+                break
+
+            nested += addr[i]
+            i += 1
+
+        assert found_paren
+
+        if addr[1] == 'A':  # ARI or ARIPI
+            assert nested[0] == 'A'
+            assert nested[1].isnumeric()
+            assert 0 <= int(nested[1]) <= 7
+
+            if i == len(addr) - 1:
+                return AssemblyParameter(EAMode.ARI, int(nested[1]))
+
+            assert addr[i + 1] == '+'
+            return AssemblyParameter(EAMode.ARIPI, int(nested[1]))
+
+        # ALA or AWA
+        assert i == len(addr) - 3
+        assert addr[len(addr) - 1] == 'W' or addr[len(addr) - 1] == 'L'
+
+        return AssemblyParameter(EAMode.AWA if addr[len(addr) - 1] == 'W' else EAMode.ALA, parse_literal(nested))
+    if addr[0] == '#':  # IMM
+        return AssemblyParameter(EAMode.IMM, parse_literal(addr[1:]))
+    if addr[0] == '-':  # ARIPD
+        assert len(addr) == 5
+        assert addr[1] == '('
+        assert addr[2] == 'A'
+        assert addr[3].isnumeric()
+        assert 0 <= int(addr[3]) <= 7
+        assert addr[4] == ')'
+
+        return AssemblyParameter(EAMode.ARIPD, int(addr[3]))
+
+    return AssemblyParameter()
+
 
 def parse_literal(literal: str):
     """
@@ -24,21 +122,22 @@ def parse_literal(literal: str):
         if len(literal) % 2 == 0:
             literal = literal[0] + '0' + literal[1:]
 
-        return bytearray.fromhex(literal[1:])
+        return int(literal[1:], 16)
     if literal[0] == '%':  # Parsing a binary literal
         assert (len(literal) - 1) % 4 == 0  # Has to be divisible by 4 to convert into hex
         hexed = hex(int(literal[1:], 2))[2:]
         if len(hexed) % 2 == 1:
             hexed = '0' + hexed
 
-        return bytearray.fromhex(hexed)
+        return int(hexed, 16)
 
     # Integer literal
+
     hexed = hex(int(literal))[2:]
     if len(hexed) % 2 == 1:
         # Odd length string, add 0 to the beginning
         hexed = '0' + hexed
-    return bytearray.fromhex(hexed)
+    return int(hexed, 16)
 
 
 def strip_comments(line: str) -> str:
