+++ tests/easier68k/core/opcodes/test_move.py
@@ -5,6 +5,7 @@ Test methods for the move command
 from easier68k.simulator.m68k import M68K
 from easier68k.core.opcodes.move import Move
 from easier68k.core.enum.ea_mode import EAMode
+from easier68k.core.models.assembly_parameter import AssemblyParameter
 from easier68k.core.enum.register import Register
 
 def test_move():
@@ -19,8 +20,8 @@ def test_move():
     # test immediate -> data register
 
     # move 123 to D2
-    src = EAMode(EAMode.IMM, 123)
-    dst = EAMode(EAMode.DRD, 2)
+    src = AssemblyParameter(EAMode.IMM, 123)
+    dst = AssemblyParameter(EAMode.DRD, 2)
 
     # make a testing move command
     mv = Move(src, dst, 'B')
@@ -36,5 +37,3 @@ def test_move_invalid_behavior():
     """
 
     a = M68K()
-
-    src = EAMode()
